# src/backend/api/v1/types/group.go
package types

import (
	"magitrickle/utils/intID"
)

type GroupsReq struct {
	Groups *[]GroupReq `json:"groups"`
}

type GroupsRes struct {
	Groups *[]GroupRes `json:"groups,omitempty"`
}

type GroupReq struct {
	ID                 *intID.ID `json:"id" example:"0a1b2c3d" swaggertype:"string"`
	Name               string    `json:"name" example:"Routing"`
	Color              string    `json:"color" example:"#ffffff"`
	Interface          string    `json:"interface" example:"nwg0"`
	FallbackInterfaces []string  `json:"fallbackInterfaces" example:"[\"wanb\"]"`
	Enable             *bool     `json:"enable" example:"true" TODO:"Make required after 1.0.0"`
	RulesReq
}

type GroupRes struct {
	ID                 intID.ID `json:"id" example:"0a1b2c3d" swaggertype:"string"`
	Name               string   `json:"name" example:"Routing"`
	Color              string   `json:"color" example:"#ffffff"`
	Interface          string   `json:"interface" example:"nwg0"`
	FallbackInterfaces []string `json:"fallbackInterfaces" example:"[\"wanb\"]"`
	Enable             bool     `json:"enable" example:"true"`
	RulesRes
}