# src/backend/api/v1/types/interface.go
package types

type InterfacesRes struct {
	Interfaces []InterfaceRes `json:"interfaces,omitempty"`
}

type InterfaceRes struct {
	ID   string `json:"id" example:"nwg0" swaggertype:"string"`
	Name string `json:"name,omitempty" example:"Home"`
	Kind string `json:"kind,omitempty" example:"interface"`
}